delete compWrapper folder, fix search
