--- src/search/compWrapper/index.js
@@ -1,3 +1,0 @@
-import SearchInputView from './SearchInputView';
-
-export { SearchInputView };
